+++ tests/benchmark.py
@@ -311,6 +311,7 @@ def _get_random_layer(cfg: ngl.SceneCfg, rng, t0, t1, enable_computes, layer=4):
             rtt = ngl.RenderToTexture(
                 child,
                 clear_color=_get_random_color(rng) + (1,),
+                forward_transforms=True,
             )
             rtt.add_color_textures(rtt_tex)
             rtt_render = ngl.RenderTexture(